+++ Magic/src/main/java/com/elmakers/mine/bukkit/world/spawn/SpawnResult.java
@@ -1,0 +1,5 @@
+package com.elmakers.mine.bukkit.world.spawn;
+
+public enum SpawnResult {
+    REPLACE, SKIP, STOP, REMOVE
+}