+++ Magic/src/main/java/com/elmakers/mine/bukkit/world/spawn/builtin/ModifyRule.java
@@ -1,11 +1,12 @@
 package com.elmakers.mine.bukkit.world.spawn.builtin;
 
-import javax.annotation.Nullable;
+import javax.annotation.Nonnull;
 
 import org.bukkit.entity.LivingEntity;
 import org.bukkit.plugin.Plugin;
 
 import com.elmakers.mine.bukkit.entity.EntityData;
+import com.elmakers.mine.bukkit.world.spawn.SpawnResult;
 import com.elmakers.mine.bukkit.world.spawn.SpawnRule;
 
 public class ModifyRule extends SpawnRule {
@@ -14,14 +15,13 @@ public class ModifyRule extends SpawnRule {
     @Override
     public void finalizeLoad(String worldName) {
         entityData = new EntityData(controller, parameters);
-        controller.info(" Modifying : " + getTargetEntityTypeName() + " at y > " + minY
-                + " at a " + (percentChance * 100) + "% chance");
+        logSpawnRule("Modifying : " + getTargetEntityTypeName() + " in " + worldName);
     }
 
     @Override
-    @Nullable
-    public LivingEntity onProcess(Plugin plugin, LivingEntity entity) {
+    @Nonnull
+    public SpawnResult onProcess(Plugin plugin, LivingEntity entity) {
         entityData.modify(entity);
-        return null;
+        return SpawnResult.SKIP;
     }
 }